Inline some small functions.

+++ cpp/mpi_utils.cpp
@@ -17,6 +17,6 @@ void mpi_utils::init(){
     std::seed_seq seq{device(),device(),device(),device()};
     // mpi_utils::prng.seed(seq);
     // mpi_utils::prng.seed(0);
-    prng.seed((size_t) std::chrono::system_clock::now().time_since_epoch().count()+proc_rank);
+    prng.seed((int) std::chrono::system_clock::now().time_since_epoch().count()+proc_rank);
     mpi_utils::root=(proc_rank==0);
 }
